refactor: 톡픽 관련 DTO 정리
